requirements2.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
pdf2image
sqlite3
googletrans
transformers
torch torchvision torchaudio